fix(modal): little cleanup
