+++ packages/react/src/components/modal/modal.tsx
@@ -1,145 +1,16 @@
-import { FunctionComponent, PropsWithChildren, ReactNode, useCallback, useEffect, useState } from 'react';
+import { FunctionComponent, PropsWithChildren, useCallback, useEffect, useState } from 'react';
 import ReactModal from 'react-modal';
-import styled, { useTheme } from 'styled-components';
+import { useTheme } from 'styled-components';
 import { useTranslation } from '../../i18n/use-translation';
-import { IconButton } from '../buttons/icon-button';
-import { DeviceContextProps, useDeviceContext } from '../device-context-provider/device-context-provider';
-
-interface StyledModalProps extends Pick<DeviceContextProps, 'breakpoints' | 'isMobile'> {
-    noPadding: boolean;
-    hasCloseButton: boolean;
-}
-
-interface ContentProps extends Pick<DeviceContextProps, 'isMobile'> {
-    noPadding: boolean;
-    hasCloseButton: boolean;
-}
-
-function getPadding({ noPadding, isMobile }: ContentProps): string {
-    if (noPadding) {
-        return '0';
-    }
-    if (isMobile) {
-        return 'var(--spacing-2x)';
-    }
-    return 'var(--spacing-4x)';
-}
-
-function getTopPadding({ hasCloseButton, noPadding, isMobile }: ContentProps): string {
-    if (noPadding) {
-        return '0';
-    }
-    if (isMobile) {
-        if (hasCloseButton) {
-            return 'var(--spacing-2x)';
-        }
-    }
-    return 'var(--spacing-3x)';
-}
-
-function getModalMinWidth({ breakpoints, isMobile }: StyledModalProps): string {
-    return isMobile ? 'initial' : `calc(${breakpoints.mobile}px - var(--spacing-4x))`;
-}
-
-const StyledModal = styled(ReactModal)<StyledModalProps>`
-    background-color: ${({ theme }) => theme.component['modal-background-color']};
-    border-radius: var(--border-radius-2x);
-    box-shadow: 0 6px 10px 0 rgb(0 0 0 / 10%);
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    max-height: calc(100vh - var(--spacing-2x));
-    max-width: 700px;
-    min-width: ${getModalMinWidth};
-    position: relative;
-    width: ${({ isMobile }) => (isMobile ? 'calc(100vw - var(--spacing-2x))' : '60vw')};
-
-    /* Firefox overflow-y: scroll problem fix (skipped bottom padding)
-    https://bugzilla.mozilla.org/show_bug.cgi?id=748518 */
-
-    &::after {
-        content: '';
-        display: block;
-        padding-bottom: ${getPadding};
-    }
-`;
-
-const Main = styled.main<ContentProps>`
-    max-height: 100%;
-    overflow-y: auto;
-    padding: ${getTopPadding} ${getPadding} 0;
-`;
-
-interface HeaderProps extends ContentProps {
-    isTopScrolled?: boolean;
-}
-const Header = styled.header<HeaderProps>`
-    border-bottom: 1px solid ${({ isTopScrolled, theme }) => (isTopScrolled ? theme.component['modal-border-color'] : 'transparent')};
-    padding: ${getTopPadding} ${getPadding} var(--spacing-2x);
-
-    & + ${Main} {
-        padding-top: 0;
-    }
-`;
-
-const CloseIconButton = styled(IconButton)<Pick<DeviceContextProps, 'isMobile'>>`
-    position: absolute;
-    right: ${({ isMobile }) => (isMobile ? 'var(--spacing-half)' : 'var(--spacing-4x)')};
-    top: 1.75rem;
-`;
-
-interface FooterProps extends ContentProps {
-    isBottomScrolled?: boolean;
-}
-const Footer = styled.footer<FooterProps>`
-    border-top: 1px solid ${({ isBottomScrolled, theme }) => (isBottomScrolled ? theme.component['modal-border-color'] : 'transparent')};
-    padding: var(--spacing-4x) ${getPadding} 0;
-`;
-
-export interface ModalProps {
-    /** Takes a query selector targeting the app Element. */
-    appElement?: string;
-    ariaDescribedby?: string;
-    /** Boolean indicating if the appElement should be hidden. Defaults to true.
-     * Should only be used for test purposes. */
-    ariaHideApp?: boolean;
-    ariaLabel?: string;
-    ariaLabelledBy?: string;
-    children?: ReactNode;
-    className?: string;
-    /**
-     * Removes padding to give you a blank modal to work with.
-     * @default false
-     */
-    noPadding?: boolean;
-    /**
-     * Adds "x" iconButton to close modal
-     * @default true
-     */
-    hasCloseButton?: boolean;
-    isOpen: boolean;
-    modalFooter?: ReactNode;
-    modalHeader?: ReactNode;
-    parentSelector?: () => HTMLElement;
-    /**
-     * Sets modal role tag
-     * @default dialog
-     */
-    role?: string;
-    /**
-     * Defines if the overlay click should close the modal
-     * @default true
-     */
-    shouldCloseOnOverlayClick?: boolean;
-
-    /** Function that will run after the modal has opened */
-    onAfterOpen?(): void;
-
-    /** Function that will run after the modal has closed */
-    onAfterClose?(): void;
-
-    onRequestClose(): void;
-}
+import { useDeviceContext } from '../device-context-provider/device-context-provider';
+import {
+    CloseIconButton,
+    Footer,
+    Header,
+    Main,
+    StyledModal,
+} from './styled';
+import { ModalProps } from './types';
 
 export const Modal: FunctionComponent<PropsWithChildren<ModalProps>> = ({
     appElement,
