+++ packages/react/src/components/modal/types.ts
@@ -1,0 +1,90 @@
+import { ReactElement, ReactNode } from 'react';
+import { DeviceContextProps } from '../device-context-provider/device-context-provider';
+import { IconName } from '../icon/icon';
+
+export interface StyledModalProps extends Pick<DeviceContextProps, 'breakpoints' | 'isMobile'> {
+    noPadding: boolean;
+    hasCloseButton: boolean;
+}
+
+export interface ContentProps extends Pick<DeviceContextProps, 'isMobile'> {
+    noPadding: boolean;
+    hasCloseButton: boolean;
+}
+
+export interface HeaderProps extends ContentProps {
+    isTopScrolled?: boolean;
+}
+
+export interface FooterProps extends ContentProps {
+    isBottomScrolled?: boolean;
+}
+
+export type MobileDeviceContextProps = Pick<DeviceContextProps, 'isMobile'>
+
+export type DialogType =
+    | 'information'
+    | 'action'
+    | 'alert';
+
+export interface BaseModalProps {
+    /** Takes a query selector targeting the app Element. */
+    appElement?: string;
+    ariaDescribedby?: string;
+    /** Boolean indicating if the appElement should be hidden. Defaults to true.
+     * Should only be used for test purposes. */
+    ariaHideApp?: boolean;
+    children?: ReactNode;
+    className?: string;
+    /**
+     * Adds "x" iconButton to close modal
+     * @default true
+     */
+    hasCloseButton?: boolean;
+    isOpen: boolean;
+    /**
+     * Defines if the overlay click should close the modal
+     * @default true
+     */
+    shouldCloseOnOverlayClick?: boolean;
+    parentSelector?: () => HTMLElement;
+
+    onRequestClose(): void;
+}
+
+export interface ModalProps extends BaseModalProps {
+    ariaLabel?: string;
+    ariaLabelledBy?: string;
+    /**
+     * Removes padding to give you a blank modal to work with.
+     * @default false
+     */
+    noPadding?: boolean;
+
+    modalHeader?: ReactNode;
+    modalFooter?: ReactNode;
+
+    /**
+     * Sets modal role tag
+     * @default dialog
+     */
+    role?: string;
+
+    /** Function that will run after the modal has opened */
+    onAfterOpen?(): void;
+
+    /** Function that will run after the modal has closed */
+    onAfterClose?(): void;
+}
+
+export interface ModalDialogProps extends BaseModalProps {
+    footerContent?: ReactElement;
+
+    title: string;
+    subtitle?: string;
+    titleIcon?: IconName;
+
+    dialogType?: DialogType;
+    confirmButton?: { label?: string, onConfirm?(): void };
+    cancelButton?: { label?: string, onCancel?(): void };
+}
